+++ core/src/main/java/zemeckis/AnimationFrameExecutor.java
@@ -1,10 +1,12 @@
 package zemeckis;
 
 import akasha.WindowGlobal;
+import grim.annotations.OmitType;
 
 /**
  * Run tasks in next AnimationFrame.
  */
+@OmitType(when = "zemeckis.use_test_scheduler")
 final class AnimationFrameExecutor extends RoundBasedExecutor {
     @Override
     void scheduleForActivation() {