+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.18](https://github.com/realityforge/zemeckis/tree/v0.18) (2026-08-05) · [Full Changelog](https://github.com/realityforge/zemeckis/compare/v0.17...v0.18)
+
+Changes in this release:
 
 * Publish release Javadocs to GitHub Pages and lint GitHub Actions workflows with Bazel.
 * Remove the dependency on Akasha by using internal browser runtime bindings.